// src/infra/interfaces/okcoin.ts
export interface Okcoin {
  best_ask: string;
  best_bid: string;
  instrument_id: string;
  open_utc0: string;
  open_utc8: string;
  product_id: string;
  last: string;
  last_qty: string;
  ask: string;
  best_ask_size: string;
  bid: string;
  best_bid_size: string;
  open_24h: string;
  high_24h: string;
  low_24h: string;
  base_volume_24h: string;
  timestamp: Date;
  quote_volume_24h: string;
}